Improve custom values querying if columns include any custom value
If any custom value is required, preload them so that the view doesn't
trigger a new SQL request per project, which can be costly. In my test
platform I have about 380 projects, which generates 380 additional
queries from the view when the projects are displayed. In that (worst
scenario) case, page rendering is 4.8s, VS 1.8s with this path.